feat: handle navigation bar insets on fab
