+++ core/scaffold/src/commonMain/kotlin/com/divinelink/core/scaffold/PersistentScaffold.kt
@@ -6,8 +6,10 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.statusBars
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
@@ -38,6 +40,7 @@ fun ScaffoldState.PersistentScaffold(
     navigationRail = navigationRail,
     content = {
       Scaffold(
+        contentWindowInsets = WindowInsets.statusBars,
         modifier = Modifier
           .animateBounds(lookaheadScope = this),
         topBar = {
